fix(player): stabilize seek pipeline and unify seek preview rendering; bump to v0.9.0
- Bump app version to 0.9.0.
- Harden Android VLC seek lifecycle in ReactVlcPlayerView:
  - add seek guards, duplicate filtering, and buffer-timeout fallback
  - improve stopped/ended seek revive path
  - suppress transient paused events during seek flush windows
  - tighten lifecycle handling for pause/resume/PiP transitions
  - add create/cleanup guards and safer teardown callback cleanup
  - dedupe onLoad metadata emission and improve media session/notification state handling
  - reuse equalizer instance instead of re-allocating on every update
- Fix native prop type consistency in ReactVlcPlayerViewManager by passing long for progress interval.
- Simplify and stabilize JS seek/progress flow in usePlayerCore:
  - clearer previewSeek/commitSeek pipeline
  - improved post-seek settlement handling
  - tighter progress/play/pause/end/seek synchronization paths
- Fix seekbar white progress disappearing during swipe seek:
  - unify seek preview source between scrubber and swipe gestures
  - wire PlayerControls to shared seek preview value
  - add scrub guard rails for invalid begin/update/finalize paths
  - add finalize safety for cancelled seek gestures

Validation:
- npx tsc --noEmit passed
- ESLint on touched files reported warnings only (no errors)

Risk notes:
- ReactVlcPlayerView is a large behavioral change; validate seek/playback, PiP, background/foreground, and notification controls on device.

+++ libs/react-native-vlc-media-player/android/src/main/java/com/yuanzhou/vlc/vlcplayer/ReactVlcPlayerViewManager.java
@@ -102,7 +102,7 @@ public class ReactVlcPlayerViewManager extends SimpleViewManager<ReactVlcPlayerV
 
     @ReactProp(name = PROP_PROGRESS_UPDATE_INTERVAL, defaultFloat = 0f)
     public void setInterval(final ReactVlcPlayerView videoView, final float interval) {
-        videoView.setmProgressUpdateInterval(interval);
+        videoView.setmProgressUpdateInterval((long) interval);
     }
 
     @ReactProp(name = PROP_PAUSED, defaultBoolean = false)
